+++ Code/LYRUIMessageCollectionViewCell.h
@@ -25,7 +25,6 @@
 @property (nonatomic) UIFont *messageTextFont UI_APPEARANCE_SELECTOR;
 @property (nonatomic) UIColor *messageTextColor UI_APPEARANCE_SELECTOR;
 @property (nonatomic) UIColor *bubbleViewColor UI_APPEARANCE_SELECTOR;
-@property (nonatomic) UIColor *pendingBubbleViewColor UI_APPEARANCE_SELECTOR;
 
 /**
  @abstract The view that handles displaying the cell content.