+++ CHANGELOG.md
@@ -10,3 +10,4 @@
 * `-[LYRUIMessageComposeTextView insertImage:]` has been removed. Use `-[LYRUIMessageInputToolbar insertImage:]` instead.
 * `-[LYRUIMessageComposeTextView removeAttachements]` has been removed. It doesn't have a replacement since it was meant for internal use only.
 * The `placeHolderText` property of `LYRUIMessageComposeTextView` is now `placeholder`.
+* The unused `pendingBubbleViewColor` property of `LYRUIMessageCollectionViewCell` has been removed.